etc: curriculums_body.dartを削除してscreenファイルにまとめた GT-3
